==> 2. Pre-processing & Feature engineering_1.ipynb <==
**1. INSTALLER LES LIBRAIRIES**

# Exécuter seulement si ce n'est pas encore intallé

# !pip install langdetect
# !pip install textblob
# !pip install -U textblob-fr

**2. IMPORTER LES LIBRAIRIES**

import pandas as pd
import numpy as np

import datetime
import re

from langdetect import detect

from textblob import Blobber
from textblob_fr import PatternTagger, PatternAnalyzer
tb = Blobber(pos_tagger=PatternTagger(), analyzer=PatternAnalyzer())

import warnings
warnings.filterwarnings('ignore')

**3. IMPORTER LE FICHIER**

# Import du fichier .csv obtenu suite au Webscrapping avec les noms de colonnes associées.
col_names = ['nom', 'notes', 'pays', 'nbr_avis', 'date_publication', 'date_experience', 'commentaire', 'titre',] 
df = pd.read_csv("1. Résultats Wepscrapping.csv", names = col_names)

df.info()
display(df.head())
print("Nombre de valeurs nulles :", df.isna().sum().sum())

**4. PRE-PROCESSING & FEATURE ENGINEERING_1**

*1) Suppression des lignes avec une valeur nulle*

df = df.dropna(axis=0, how='any')
# Vérification
print("Nombre de valeurs nulles restant :", df.isna().sum().sum())

*2) Conserver uniquement les avis provenant de France*

print("Nombre d'avis provenant d'un pays autre que la France :", len(df.loc[(df['pays'] != 'FR')]))
df = df.drop(df[(df['pays'] != 'FR')].index)
# Réinitialiser l'index
df = df.reset_index(drop=True)
# Vérification
print("Nombre d'avis restant provenant d'un pays autre que la France :", len(df.loc[(df['pays'] != 'FR')]))

*3) Fusionner les colonnes 'commentaire' et 'titre'*

# Nous avons déjà supprimer les lignes avec une valeur nulle.
# Ici, les lignes sans commentaires sont indiquées par les termes "No comment" dans la colonne "commentaire".
print("Nombre de lignes sans commentaires :", df.loc[(df['commentaire'] == 'No comment')].shape[0])

# Fusionner les colonnes 'commentaires' et 'titre' dans une nouvelle colonne 'commentaires'
commentaires = []
for i in df.index :
    # S'il n'y pas de commentaires, définir le titre comme le commentaire
    if (df['commentaire'][i] == "No comment") :
        commentaires.append(df['titre'][i])
    # Si le titre et le commentaire sont identiques, conserver uniquement le commentaire
    elif (df['titre'][i] == df['commentaire'][i]):
        commentaires.append(df['commentaire'][i])
    # Sinon, fusionner le titre et le commentaire
    else : 
        commentaires.append(df['titre'][i] + " " + df['commentaire'][i])

# Ajouter la nouvelle colonne 'commentaires' au Dataframe
nouvelle_colonne = pd.DataFrame(commentaires, columns=["commentaires"])
df = pd.concat([df, nouvelle_colonne], axis = 1)

# Vérification
df.head()

*4) Conserver uniquement les commentaires en français*

# Certains commentaires provenant de France sont dans une autre langue, nous les supprimons :
for i in df.index :
  try :
      if detect(df["commentaires"][i]) != "fr" :
          df = df.drop(index = i)
  except : 
      df = df.drop(index = i)

# Réinitialiser l'index
df = df.reset_index(drop=True)

*5) Transformer la colonne 'nbr_avis' et la mettre au bon format*

# '3 reviews' (object) >>> 3 (int64)
df['nbr_avis'] = df['nbr_avis'].apply(lambda x : x.split(' ')[0])
df['nbr_avis'] = df['nbr_avis'].astype('int64')
# Vérification
print("Format de la colonne 'nbr_avis' :", df['nbr_avis'].dtype)

*6) Mettre la colonnes 'date_experience' au format datetime*

df['date_experience'] = df['date_experience'].apply(lambda date: date.split(': ')[1])
df['date_experience'] = pd.to_datetime(df['date_experience'])
# Vérification
print("Format de la colonne 'date_experience' :", df['date_experience'].dtype)

*7) Création de la variable 'ponctuation'*

# La variable "ponctuation" comptabilise le nombre de '!', '?' et '...' dans un commentaire :
def compteur_ponctuation(str):
    r = re.compile("[!\?]") 
    r1 = re.compile("[...]{3,}") 
    str1 = r.findall(str)
    str2 = r1.findall(str)
    return len(str1)+len(str2)
    
# Ajout de la colonne dans le DataFrame
df['ponctuation'] = df['commentaires'].apply(lambda x : compteur_ponctuation(x))

*8) Création de la variable 'sentiment'*

# Variable quantitative
def sentiment(commentaire):
    blob = tb(commentaire)
    return blob.sentiment[0]

# Ajout de la colonne dans le DataFrame
df['sentiment'] = df['commentaires'].apply(lambda x : sentiment(x))

*9) Création de la variable 'sentiment_labels'*

# Variable qualitative : Classer les sentiments selon les labels 'Positif', 'Negatif' ou 'Neutre'
senti_list = []
for i in range(len(df['sentiment'])):
    if (df['sentiment'][i] > 0):
        senti_list.append('Positif')
    elif (df['sentiment'][i] < 0):
        senti_list.append('Negatif')
    else:
        senti_list.append('Neutre')

# Ajout de la colonne dans le DataFrame
df["sentiment_labels"] = senti_list

*10) Vérification finale*

df.info()
df.head()

**5. DATAFRAME DANS UN FICHIER .CSV**

df.to_csv("2. Pre-processing & Feature engineering_1.csv", index=False)
