==> 6.b) Modélisation_1 et 2 (AdaBoost_stemming).ipynb <==
**1. IMPORTER LES LIBRAIRIES**

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline 
import seaborn as sns

# Train_Test_Split
from sklearn.model_selection import train_test_split

# Ré-échantillonnage
from imblearn.over_sampling import RandomOverSampler
from imblearn.under_sampling import RandomUnderSampler

# Modélisation
from sklearn.ensemble import AdaBoostClassifier
from sklearn.model_selection import RandomizedSearchCV

# Performances
from imblearn.metrics import classification_report_imbalanced
from sklearn.metrics import f1_score

# Enregistrer le modèle
from sklearn import model_selection
import joblib

import warnings
warnings.filterwarnings('ignore')

**2. IMPORTER LE FICHIER**

*Stemming / TF-IDF / Réduction de dimension*

df_stem = pd.read_csv("4. Stemming_TF-IDF(Avec_Reduction).csv").astype('float16')
df_stem['y'] = df_stem['y'].astype('uint8')
df_stem = df_stem.rename(columns={'1099' : 'ponctuation', '1100' : 'sentiment'})

df_stem.info()
display(df_stem.head())

# **3. MODELISATION_1 (5 CLASSES)**

*1) Préparation des données*

# Target / Features
X_stem = df_stem.drop('y', axis = 1).to_numpy()
y_stem = df_stem['y'].to_numpy()

# Train_Test_Split
X_train_stem, X_test_stem, y_train_stem, y_test_stem = model_selection.train_test_split(X_stem, y_stem, test_size=0.2, random_state = 42)

# Vérification shape
print ("X_train_stem =", X_train_stem.shape)
print ("y_train_stem =", y_train_stem.shape)
print ("X_test_stem =", X_test_stem.shape)
print ("y_test_stem =", y_test_stem.shape)

# Undersampling
rUs = RandomUnderSampler(random_state = 42)
X_ru_stem, y_ru_stem = rUs.fit_resample(X_train_stem, y_train_stem)
print('Classes échantillon undersampled stemming :', dict(pd.Series(y_ru_stem).value_counts()))

*2) Modélisation*

adaboost_model = AdaBoostClassifier(random_state = 42)
adaboost_params = {'n_estimators': [10, 50, 100], # Nombre d'estimateurs dans l'ensemble
                   'learning_rate': [0.01, 0.1, 0.5, 1]} # Taux d'apprentissage du modèle

# Création de l'objet RandomizedSearchCV pour trouver les meilleurs hyperparamètres
adaboost_random_1 = RandomizedSearchCV(adaboost_model, adaboost_params, n_iter=4, cv=5, n_jobs=-1)
adaboost_random_1.fit(X_ru_stem, y_ru_stem)
# Sélection du meilleur modèle trouvé par RandomizedSearchCV
best_adaboost_model_1 = adaboost_random_1.best_estimator_

*3) Performances*

y_pred_stem = best_adaboost_model_1.predict(X_test_stem)

print("Score du modèle :", best_adaboost_model_1.score(X_test_stem, y_test_stem))

print(classification_report_imbalanced(y_test_stem, y_pred_stem))
      
confusion_matrix = pd.crosstab(y_test_stem, y_pred_stem, rownames=['Classe réelle'], colnames=['Classe prédite'])
display(confusion_matrix)

# f1_score du modèle AdaBoost_1
print('f1_score AdaBoost_1 =', f1_score(y_test_stem, y_pred_stem, average='weighted'))

# f1_score par classe du modèle AdaBoost_1
AD_f1_stem_1 = f1_score(y_test_stem, y_pred_stem, average=None)
print('f1_score par classes AdaBoost_1 =', AD_f1_stem_1)

*4) Enregistrement du modèle*

joblib.dump(best_adaboost_model_1, '6.b) adaboost_1.pkl')

# **4. MODELISATION_2 (BINAIRE)**

*1) Préparation des données*

### Classification binaire
# Classe 0 : [1, 2, 3]
# Classe 1 : [4, 5]
df_stem['y'] = df_stem['y'].replace((1, 2, 3, 4, 5), (0, 0, 0, 1, 1))

# Target / Features
X_stem = df_stem.drop('y', axis = 1).to_numpy()
y_stem = df_stem['y'].to_numpy()

# Train_Test_Split
X_train_stem, X_test_stem, y_train_stem, y_test_stem = model_selection.train_test_split(X_stem, y_stem, test_size=0.2, random_state = 42)

# Undersampling
rUs = RandomUnderSampler(random_state = 42)
X_ru_stem, y_ru_stem = rUs.fit_resample(X_train_stem, y_train_stem)
print('Classes échantillon undersampled stemming :', dict(pd.Series(y_ru_stem).value_counts()))

*2) Modélisation*

adaboost_model = AdaBoostClassifier(random_state = 42)
adaboost_params = {'n_estimators': [10, 50, 100], # Nombre d'estimateurs dans l'ensemble
                   'learning_rate': [0.01, 0.1, 0.5, 1]} # Taux d'apprentissage du modèle

# Création de l'objet RandomizedSearchCV pour trouver les meilleurs hyperparamètres
adaboost_random_2 = RandomizedSearchCV(adaboost_model, adaboost_params, n_iter=4, cv=5, n_jobs=-1)
adaboost_random_2.fit(X_ru_stem, y_ru_stem)
# Sélection du meilleur modèle trouvé par RandomizedSearchCV
best_adaboost_model_2 = adaboost_random_2.best_estimator_

*3) Performances*

y_pred_stem = best_adaboost_model_2.predict(X_test_stem)

print("Score du modèle :", best_adaboost_model_2.score(X_test_stem, y_test_stem))

print(classification_report_imbalanced(y_test_stem, y_pred_stem))
      
confusion_matrix = pd.crosstab(y_test_stem, y_pred_stem, rownames=['Classe réelle'], colnames=['Classe prédite'])
display(confusion_matrix)

# f1_score du modèle AdaBoost_2
print('f1_score AdaBoost_2 =', f1_score(y_test_stem, y_pred_stem, average='weighted'))

# f1_score par classe du modèle AdaBoost_2
AD_f1_stem_2 = f1_score(y_test_stem, y_pred_stem, average=None)
print('f1_score par classes AdaBoost_2 =', AD_f1_stem_2)

*4) Enregistrement du modèle*

joblib.dump(best_adaboost_model_2, '6.b) adaboost_2.pkl')
